--- a/ecc/bls24-317/fr/asm_avx.go
+++ b/ecc/bls24-317/fr/asm_avx.go
@@ -10,6 +10,6 @@ package fr
 import "golang.org/x/sys/cpu"
 
 var (
-	supportAvx512 = supportAdx && cpu.X86.HasAVX512 && cpu.X86.HasAVX512DQ
+	supportAvx512 = supportAdx && cpu.X86.HasAVX512 && cpu.X86.HasAVX512DQ && cpu.X86.HasAVX512VBMI2
 	_             = supportAvx512
 )