--- a/field/babybear/fft/kernel_amd64.go
+++ b/field/babybear/fft/kernel_amd64.go
@@ -29,12 +29,23 @@ var vInterleaveIndices = []uint64{
 //go:noescape
 func innerDIFWithTwiddles_avx512(a []babybear.Element, twiddles []babybear.Element, start, end, m int)
 
+//go:noescape
+func innerDITWithTwiddles_avx512(a []babybear.Element, twiddles []babybear.Element, start, end, m int)
+
 func innerDIFWithTwiddles(a []babybear.Element, twiddles []babybear.Element, start, end, m int) {
 	if !supportAVX512 {
 		innerDIFWithTwiddlesGeneric(a, twiddles, start, end, m)
 		return
 	}
 	innerDIFWithTwiddles_avx512(a, twiddles, start, end, m)
+}
+
+func innerDITWithTwiddles(a []babybear.Element, twiddles []babybear.Element, start, end, m int) {
+	if !supportAVX512 {
+		innerDITWithTwiddlesGeneric(a, twiddles, start, end, m)
+		return
+	}
+	innerDITWithTwiddles_avx512(a, twiddles, start, end, m)
 }
 
 //go:noescape