--- a/field/generator/internal/addchain/addchain.go
+++ b/field/generator/internal/addchain/addchain.go
@@ -64,7 +64,6 @@ var (
 
 // GetAddChain returns template data of a short addition chain for given big.Int
 func GetAddChain(n *big.Int) *AddChainData {
-
 	// init the cache only once.
 	once.Do(initCache)