--- a/field/generator/internal/templates/element/vector_ops_asm.go
+++ b/field/generator/internal/templates/element/vector_ops_asm.go
@@ -142,6 +142,9 @@ var (
 //go:noescape
 func mulVec(res, a, b *{{.ElementName}}, n uint64, qInvNeg uint64)
 
+
+
+
 `
 
 const VectorOpsArm64 = VectorOpsPureGo
@@ -264,6 +267,9 @@ func scalarMulVec(res, a, b *{{.ElementName}}, n uint64)
 
 //go:noescape
 func innerProdVec(t *uint64, a, b *{{.ElementName}}, n uint64)
+
+//go:noescape
+func butterflyMulVec(a, twiddles *{{.ElementName}}, m int)
 
 // Add adds two vectors element-wise and stores the result in self.
 // It panics if the vectors don't have the same length.
@@ -428,7 +434,4 @@ func (vector *Vector) Mul(a, b Vector) {
 		mulVecGeneric((*vector)[start:], a[start:], b[start:])
 	}
 }
-
-
-
 `