--- a/field/generator/internal/templates/element/vector_ops_purego.go
+++ b/field/generator/internal/templates/element/vector_ops_purego.go
@@ -37,4 +37,5 @@ func (vector *Vector) InnerProduct(other Vector) (res {{.ElementName}}) {
 func (vector *Vector) Mul(a, b Vector) {
 	mulVecGeneric(*vector, a, b)
 }
+
 `