… ecc/bls12-377/fr/fft/fft_test.go
… ecc/bls12-377/fr/sis/sis_test.go
… ecc/bls12-381/fr/fft/fft_test.go
… ecc/bls24-315/fr/fft/fft_test.go
… ecc/bls24-317/fr/fft/fft_test.go
… ecc/bn254/fr/fft/fft_test.go
… ecc/bw6-633/fr/fft/fft_test.go
… ecc/bw6-761/fr/fft/fft_test.go
… field/babybear/element_amd64.s
… field/babybear/fft/fft_test.go
… field/babybear/sis/sis_test.go
… field/generator/generator_fft.go
… field/generator/internal/templates/fft/fft.go.tmpl
… field/generator/internal/templates/fft/kernel.purego.go.tmpl
… field/generator/internal/templates/fft/tests/fft.go.tmpl
… field/generator/internal/templates/sis/sis.go.tmpl
… field/generator/internal/templates/sis/sis.test.go.tmpl
… field/goldilocks/fft/fft.go
… field/goldilocks/fft/fft_test.go
… field/goldilocks/fft/kernel_purego.go
… field/goldilocks/sis/sis.go
… field/goldilocks/sis/sis_test.go
… field/internal/main.go
… field/koalabear/asm_avx.go
… field/koalabear/element_amd64.s
… field/koalabear/fft/fft.go
… field/koalabear/fft/fft_test.go
… field/koalabear/fft/kernel_amd64.go
… field/koalabear/fft/kernel_amd64.s
… field/koalabear/fft/kernel_purego.go