# Description

This PR adds optimized (AVX512 assembly) FFT kernels (DIF only for now) for `koalabear` and `babybear` fields.
This is a pre-requisite for faster SIS hashing, which will have its own dedicated PR.

Still a bit of room for reducing these numbers; right now kernel of size 128 is fully in AVX512, but we could extend that a bit to do the full 512 size (matching SIS hashing params) if needed.

## Benchmark on `koalabear`

FFT DIF with domain cardinality == 512:

```
BenchmarkFFTDIFReferenceSmall-16     11604         2352          -79.73%
BenchmarkFFTDIFReferenceSmall-16     11668         2398          -79.45%
BenchmarkFFTDIFReferenceSmall-16     12207         2280          -81.32%
```

With cardinality == 1 <<20
```
BenchmarkFFTDIFReference-16     2950281       934681        -68.32%
BenchmarkFFTDIFReference-16     3422121       937623        -72.60%
BenchmarkFFTDIFReference-16     3442894       938829        -72.73%
```

RingSIS with log2(bound) == 16 and degree == 512 (secure parameters used in linea prover)

```
benchmark                                                             old ns/op     new ns/op     delta
BenchmarkSIS/ring-sis/inputs=65536/log2-bound=16/log2-degree=9-16     3315941       407467        -87.71%
BenchmarkSIS/ring-sis/inputs=65536/log2-bound=16/log2-degree=9-16     3306674       407898        -87.66%
BenchmarkSIS/ring-sis/inputs=65536/log2-bound=16/log2-degree=9-16     3296878       409011        -87.59%

benchmark                                                             old allocs     new allocs     delta
BenchmarkSIS/ring-sis/inputs=65536/log2-bound=16/log2-degree=9-16     260            3              -98.85%
BenchmarkSIS/ring-sis/inputs=65536/log2-bound=16/log2-degree=9-16     260            3              -98.85%
BenchmarkSIS/ring-sis/inputs=65536/log2-bound=16/log2-degree=9-16     260            3              -98.85%

benchmark                                                             old bytes     new bytes     delta
BenchmarkSIS/ring-sis/inputs=65536/log2-bound=16/log2-degree=9-16     14464         4144          -71.35%
BenchmarkSIS/ring-sis/inputs=65536/log2-bound=16/log2-degree=9-16     14464         4144          -71.35%
BenchmarkSIS/ring-sis/inputs=65536/log2-bound=16/log2-degree=9-16     14464         4144          -71.35%
```